Fix issue with wrong values showing up on BBB boot pins during power on sequence
Set the values of the pin drives during the boot sequence of the BBB and Cypress devices so that bad values won't end up on those pins and screw up the boot configuration.  This would cause corrupt RS-232 console communiucation and/or prevent booting of certain types of BBB board (the black ones).